debug

+++ SocialNetwork/resources/views/newContent.view.php
@@ -32,46 +32,48 @@ if(array_key_exists('content', $_POST) && !array_key_exists('entry', $_GET)) {
 
 //if(isset($_POST['picture'])) {
 	if(isset($_FILES['picture']['name'])) {
-		$upload_folder = 'img/content/posts/'. $username .'/'; //Das Upload-Verzeichnis
-		if(!file_exists($upload_folder)) {
-			mkdir($upload_folder);
-		}
+		if($_FILES['picture']['name'] != "") {
+			$upload_folder = 'img/content/posts/'. $username .'/'; //Das Upload-Verzeichnis
+			if(!file_exists($upload_folder)) {
+				mkdir($upload_folder);
+			}
 
-		$filename = pathinfo($_FILES['picture']['name'], PATHINFO_FILENAME);
-		$extension = strtolower(pathinfo($_FILES['picture']['name'], PATHINFO_EXTENSION));
-		 
-		 
-		//Überprüfung der Dateiendung
-		$allowed_extensions = array('png', 'jpg', 'jpeg', 'gif');
-		if(!in_array($extension, $allowed_extensions)) {
-		 	die("Ungültige Dateiendung. Nur png, jpg, jpeg und gif-Dateien sind erlaubt");
+			$filename = pathinfo($_FILES['picture']['name'], PATHINFO_FILENAME);
+			$extension = strtolower(pathinfo($_FILES['picture']['name'], PATHINFO_EXTENSION));
+			 
+			 
+			//Überprüfung der Dateiendung
+			$allowed_extensions = array('png', 'jpg', 'jpeg', 'gif');
+			if(!in_array($extension, $allowed_extensions)) {
+			 	die("Ungültige Dateiendung. Nur png, jpg, jpeg und gif-Dateien sind erlaubt");
+			}
+			 
+			//Überprüfung der Dateigröße
+			$max_size = 500*1024; //500 KB
+			if($_FILES['picture']['size'] > $max_size) {
+			 	die("Bitte keine Dateien größer 500kb hochladen");
+			}
+			 
+			//Überprüfung dass das Bild keine Fehler enthält
+			if(function_exists('exif_imagetype')) { //Die exif_imagetype-Funktion erfordert die exif-Erweiterung auf dem Server
+			 	$allowed_types = array(IMAGETYPE_PNG, IMAGETYPE_JPEG, IMAGETYPE_GIF);
+			 	$detected_type = exif_imagetype($_FILES['picture']['tmp_name']);
+			 	if(!in_array($detected_type, $allowed_types)) {
+			 		die("Nur der Upload von Bilddateien ist gestattet");
+			 	}
+			}
+			 
+			//Pfad zum Upload
+			$new_path = $upload_folder.$contentId.'.'.$extension;
+			 
+			//Neuer Dateiname falls die Datei bereits existiert
+			if(file_exists($new_path)) { //Falls Datei existiert, lösche die alte Datei
+			 	unlink($new_path);
+			}
+			 
+			//Alles okay, verschiebe Datei an neuen Pfad
+			move_uploaded_file($_FILES['picture']['tmp_name'], $new_path);
 		}
-		 
-		//Überprüfung der Dateigröße
-		$max_size = 500*1024; //500 KB
-		if($_FILES['picture']['size'] > $max_size) {
-		 	die("Bitte keine Dateien größer 500kb hochladen");
-		}
-		 
-		//Überprüfung dass das Bild keine Fehler enthält
-		if(function_exists('exif_imagetype')) { //Die exif_imagetype-Funktion erfordert die exif-Erweiterung auf dem Server
-		 	$allowed_types = array(IMAGETYPE_PNG, IMAGETYPE_JPEG, IMAGETYPE_GIF);
-		 	$detected_type = exif_imagetype($_FILES['picture']['tmp_name']);
-		 	if(!in_array($detected_type, $allowed_types)) {
-		 		die("Nur der Upload von Bilddateien ist gestattet");
-		 	}
-		}
-		 
-		//Pfad zum Upload
-		$new_path = $upload_folder.$contentId.'.'.$extension;
-		 
-		//Neuer Dateiname falls die Datei bereits existiert
-		if(file_exists($new_path)) { //Falls Datei existiert, lösche die alte Datei
-		 	unlink($new_path);
-		}
-		 
-		//Alles okay, verschiebe Datei an neuen Pfad
-		move_uploaded_file($_FILES['picture']['tmp_name'], $new_path);
 	}
 //}
